custom-elements/creating-and-passing-registries/existing-registry-test-004.html

<!DOCTYPE html>
<html>
<head>
<title>New document without browsing context must not have a registry.</title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="In all other cases, new documents must not have a registry.">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#creating-and-passing-registries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    assert_throws('NotSupportedError',
                  function() { doc.register('x-element'); },
                  'Registering valid custom element in a document ' +
                  'without browsing context should fail');

}, 'Document without browsing context must not have a registry');


</script>
</body>
</html>
